Making the bot active - Push Messaging (#22)
* Added active messaging system into the library

* Reformatting, cleanup and documentation

* Updated README

* Fixing permissions

* Updating the active messaging push method to not send message when channel name is not available

Earlier, the method would have tried to send the message even if the channel name was not there.
This could cause the send method to fail. Also, there was no way for the pushNotifier function (closure)
to return a value which would cause a method to NOT be sent. Now there is! This is useful in case the
pushNotifier method is run but for some reason the method determines that there is no message to be
sent right now.